Revert "Этап 2. Основной сервис"
This reverts commit d37487e7

+++ stats/stats-dto/src/main/java/ru/practicum/hit/dto/HitDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class HitDto {
     String app;
     String uri;
-    Integer hits;
+    Long hits;
 }
